requirements.txt
anthropic==0.25.0
langchain==0.1.14
langchain-core==0.1.31
langchain-anthropic==0.1.8
langchain-google-genai==0.0.8
langchain-gmail==0.1.0
langgraph==0.0.49
python-telegram-bot==20.3
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
apscheduler==3.10.4
pydantic==2.5.0
requests==2.31.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.107.0
google-generativeai==0.3.0
